web_server: Allow the old default realm if no specific realm is passed.
WARP firmwares up to 1.2.4 used the default realm of the web server
"asyncesp". When updating to >= 1.3.0, some browsers don't realize
that the realm in the HTTP response is changed and continue to try
to authenticate with the old realm. If no specific realm is passed
to checkDigestAuthentication, we now accept the old "asyncesp" as
well as the new default "esp32-lib".