config: Add ConfigRoot. Move validation. Add minChars to ConfString.
Configs are now always validated with a default validator. Custom
validation is done by adding the validator to a ConfigRoot object.
The API layer now expects ConfigRoot pointers, as the root object
is the only one to support updating from JSON and ConfUpdates.

The default validator now also checks strings for a minimum length.

Note that this is a breaking change, as the parameter order (and
the default parameters) will now interpret Conf::Str("", 5) as a
string with a minimum length of 5, not (as before) a maximum length
of 5.