config: Allow passing integer values for floats.
The check was intended to make sure no additional rounding happens.
However the first int to be rounded is 2^24+1 = 16777217. This is
probably too large to be of any use of (writing!) APIs.

Also as this is JSON, a User could still send 16777217.0 as float and
have the rounding happen anyway.