web_sockets: Fix connection loss after ~ 20 seconds. Fixes #33.
The issue was that esp-idf for some reason handles ping frames
by itself even if we set handle_ws_control_frames = true.

When calling https_ws_get_frame_type the idf receives the whole
frame and sends a pong back. However after this, our handler is
still called with nothing to receive anymore. In this case we
can't even detect, that a ping frame was handled, as we would
have to call httpd_ws_recv_frame for that and the buffer is already
empty.

The fix is to apply the commited patch to the idf when rebuilding
the arduino-esp32 libs. This patch removes the broken handling of
ping frames, enabling us to handle those ourselves.