Add initial MVP firmware for Energy Manager
The WARP Energy Manager consists of a ESP32 Ethernet together with the
WARP Energy Manager Bricklet.

Initially it will have three energy manager specific frontend modules:

* Energy Manager (for the configuration of the actual manager)
* EM Meter Config (for the configuration of the meter that is used)
* EM Energy Analysis (for pretty graphs und CSV downloads)

In the backend there is an additional EM Meter module that provides the
same API as the EVSE V2 Meter module from WARP2.

A minimal UI for configuring the meter and the energy manager
is implemented. Later on there will be a "wallbox configuration"
UI that is similar to the charge manager in WARP2. Currently we
use the charge manager as is.

TODO for MVP:
* Energy Analysis front- and backend
* Energy Manager backend
* Charge Manager -> Wallbox Configuration adaptation
* Status page integration for all new frontend modules
* Translations to english