evse_v2_meter: Read meter values once in setup.
This fixes a race condition where the users module in its setup
tracks a charge stop if the ESP rebooted while a charge stopped.
The tracked stop attempts to read the meter value, first checking
if meter/state["state"] is 2.

The state is set to 2 in setup, however the first meter values
were read in a scheduled task, i.e. _after_ all module setups
are executed.