Fix digest cache invalidation
Moving the digest files to the environment specific build directory
to allow the clean target to delete them resulted in decoupling the
digest files from the build results. This broke the digest cache
invalidation in a way that made the cache assume the build results
are up-to-date while in reality the build result are up-to-date but
for a different environment.

Store a copy of the digest along the build result and check that it
is in sync with the digest stored in the build directory.