modbus_reader: Add work-around for SDM72DM reset response bug.
The SDM72DM has a bug in it's modbus implementation where it
responds to "write multiple registers"-Requests that trigger
a reset of the relative energy value with the wrong DeviceID.
It sends ID 0 which is reserved for broadcasts and never to
be used by slaves!

This only happens when writing to 461457,i.e. the reset trigger
register. In the future we should check that the reset worked
by re-reading the energy value, making sure that it is a small
enough value and retrying the reset if not.