web_sockets: Protect aginst control socket close.
This seems to happen in at least one specific setup where the web
interface (and with it WebSocket connections) is active over multiple
days on a router that drops wifi connections every few hours.

If the control socket is closed we don't have any chance to send
WebSocket frames anymore. A possible solution would be to restart
the web server, but then we would have to re-register all URL
handlers, i.e. call register_urls on every module, which probably
creates more new problems than it solves.

As this is a really esoteric case, we handle this by just restarting
the ESP completely. To protect against boot looping, we only allow
restarts after one hour and only if starting the worker fails
for a minute (~ 600 attempts to start the worker).