charge_tracker: Don't set Content-Length for chunked response.
https://httpwg.org/specs/rfc7230.html#header.content-length
says "A sender MUST NOT send a Content-Length header field
in any message that contains a Transfer-Encoding header field."

We use Transfer-Encoding: Chunked as an easy way to not hold the
complete charge log in the RAM at once.

Setting the content length is understood by browsers, resulting
in a correct progress bar for the download. However the charge log
is downloaded by the web interface's javascript code anyway
(and then converted to CSV), so the progress bar is not
visible in the common case.