api: Add get_maybe and save_maybe.
Typescript has no conditional compilation. However we want to be able
to write modules that use the API of other modules, if they are
available. To do this, use get_maybe (returns null if module is not
available) and save_maybe (does nothing if module is not available).

Currently there is no good solution to register an event handler
that will never trigger if a module is not available.
Use source.addEventListener('missing/module' as any,...) for now.