rtc: Remove centiseconds from rtc/state.
We still keep the centiseconds in rtc/state_update, to set the RTC as
accurately as possible. However in the time read from the RTC,
communicating the centiseconds is not useful, as it is
a) only refreshed every 250 ms (so the info is stale) and
b) because of that the API is sent 4 times per second.