rtc: Rename rtc/state to rtc/time. Update every 200 ms.
The RTC Bricklet reads the current time every 250 ms. If we read the
time every 200 ms (and also send it out if it changes every 100 ms),
it is fast in the web interface that the display seems to be more or
less in sync with the users system clock. Also we don't miss any updates.