task_scheduler: Fix tasks not being executed after 2^32 ms.
The task scheduler keeps a priority queue of tasks to execute. Tasks
are sorted by their next deadline, so that we only have to check the
first (in the queue) task's deadline in the loop function.

If the first task deadline wraps around, it was previously sorted at
the start of the queue, so all other tasks (with comparatively large
deadlines) won't be executed any more, even though their deadline is
elapsing before the overflow.

We fix this by first checking if we are "close" to the overflow i.e.
on the second half of the uint32_t ring. If this is the case, a task
that is in the first half of the ring (closer to 0 than to 2^32 - 1)
is regarded as less prioritized than a task closer to 2^32 - 1. This
works because we assume that all task deadlines are in the future as
we only run tasks with a fixed delay (not with a fixed interval).