Add watchdog module.
Espressif's task watchdog only allows monitoring FreeRTOS tasks.
However we want several conditions to be monitored, that are not
necessarily tied to a specific task. Those conditions can be monitored
with this watchdog module.

The watchdog module is running as a separate FreeRTOS task
and is itself monitored by the task watchdog.