config, modules: Start moving away from unsafe uses of config objects' asCString().
asCString() returns a pointer to the internal char buffer of String objects,
which is not guaranteed to be valid beyond the scope in which asCString() was
called. Code must not store this pointer but the compiler cannot enforce that.
Introduce a new function asEphemeralCString() to make the ephemeral nature of
the returned pointer more obvious to the programmer.

Safe uses of asCString() have been moved to asEphemeralCString(), but several
possibly unsafe uses remain.