api, charge_manager: Allow circumventing typing for event listeners.
The addEventListener() is intentionally strongly typed to catch non-existing
events at compile time. Therefore, it is impossible to compile code that
includes events for modules that are not going to be compiled in. This adds
an intentionally untyped variant of the function to allow compilation with
disabled modules, and comments the use of this function.