input_date: Handle timezones correctly.
User selected dates are in the browsers timezone. By using the
Date constructor with split parameters, the Date object used as
in value/onValue is in UTC with the correct offset.

So if for example the User selects the 1st of January 2023, the
resulting Date object is equivalent to the unix timestamp 1672527600
(2022-12-31T23:00:00Z)