util: Add allow_render. Use in components.
(Almost) every component has state that is set via the web socket
connection. Components did check in various ways if their state is set
and only then render.

To simplify this, add util.allow_render, a bool that is only set to
true when the first web socket message is processed completely. The
first message always contains the complete state of the charger, i.e.
all component states that depend on the web socket connection should
then be initialized.