Add support for on upload conversion of hex GD firmwares
This change adds support for uploading GD firmware files in
intel hex format just as the vendor distributes those files.
The file is converted on the fly into binary and stored on the
LittleFS. Right after uploading the firmware will automatically
be flashed to the GD just like thie binary firmware.