charge_tracker: Don't show negative charged kWh.
In some cases (for a currently unknown reason!) a charge is tracked
that has its meter_end value set to 0. This should not happen, as
we write NAN if the energy meter is not available when a charge stops.
However this has happened in the wild at least once, so we have to
handle this case.

Report such charges as if the meter was not available.