util.py: Support firmwares with beta suffix. Patch version number.
Embedded beta firmwares have their major version set to
200 + beta_version. For example
"bricklet_evse_v2_firmware_2_1_12_beta1.zbin" will be patched to
201.1.12.

As the version number changes with every beta version, the firmware
is reflashed automatically. Also we now see if the firmware checked
in the repo is a beta or a released firmware.