ethernet: Apply IP config inside event handler.
Apply IP config inside the ETH_CONNECTED event handler for a more reliable
IP configuration, especially with static IP configurations. Setting the IP
config anywhere else is subject to a race condition and often causes extra
DHCP requests, even with a static IP configuration, and may lead to a loss
of connection when the ethernet cable is reconnected.

The DHCP client default should be changed to disabled in the Arduino ESP32
framework as well.