energy_manager, evse, evse_v2, nfc: Register APIs if device not found.
This makes sure all APIs are always registered, even though the
corresponding device is missing. This simplifies developing the
web interface, because all APIs have "sane" default values.
Feature checks are only necessary if we want to write to an API
and the possible misuse should not be fatal, as all embedded C
bindings functions check whether the device is initialized before
doing anything.