config: Always require ConfStr min/max lengths.
Before this change the ::Str() constructor used the length of the
default value as max length if none was given. This allowed misusing
the constructor with ::Str(""), which constructs a ConfStr that can
only hold the empty string.

If a string longer than the allowed maximum is put into a config
(updateString does not rerun validators!) Config::toString will drop
other config members if ArduinoJson's memory buffer is full.