http: Fix race condition between uri matcher and handler registration.
httpd_register_uri_handler is not thread-safe. If a request is handled
while registering a new handler, it can happen that the uri matcher is
called with nullptr as ref_uri. Return false in case any of the URIs
are null or the ref_uri has length 0. Also add boot_stage and don't
match any URIs if the register_url-stage is not done.