config: Lock update_from_visitor.
If multiple APIs attempt to update (different!) configs at the same time
both copy their target config into a temporary. This allocates and frees
config value slots. With the right timing, a slot can then be allocated
for two config values. When the temporary copy is destroyed, the slot
is freed two times.

Locking update_from_visitor is sufficient, as this is the only point
where slots are allocated/freed once the setup is done. All setup code
runs in the main thread, before other threads are allowed to access the
config.