pio_hooks: Support fragments and functions in translations.
Translations can now contain
- strings (as before)
- fragments (for example: <>see <a href="/">here</a></>)
- and functions that return strings or fragments

Functions must only take primitive arguments (f.e. string, number, etc)
and return a string or fragment.

To add a fragment or function to a translation:
- Change the translation file from .json to .tsx
- Add the following header:
  /** @jsxImportSource preact */
  import { h } from "preact";
  let x =
  before the first {
- write fragments as expected: "test": <>blah <b>bold</b></>
- write functions inside marker comments:
  "test": /*SFN*/(x: number,) => "max" + x + "chargers"/*NF*/

Allowed markers are
- /*SFN*/ to start a function that returns a string
- /*FFN*/ to start a function that returns a fragment
- /*NF*/ to end any function (as in FN reversed or eNd Function)

Translations are type-checked, so the same translation value has to
have the same type in all languages. Functions are strongly typed too.

Because of this, typescript will complain if you try to use
a translation value that is not a (function returning a) string but
a fragment in a context that requires strings only. For example setting
the value of an input.

With this change we should be able to eliminate:
- usages of dangerouslySetInnerHTML by using fragments instead
- constructs with _prefix, _infix and _suffix placeholders
  by using functions instead