task_scheduler: Add task IDs and TaskScheduler::cancel().
A task can be cancelled by ID if it is still in the task queue.
A task that is currently being executed can not be interrupted.
However if a (repeated) task is cancelled while it is executed,
it will not be put back into the task queue when it is done.

Also add TaskScheduler::currentTask() to get the task ID of the
task that is currently being executed. This method enforces that
the calling thread is the main thread. This allows to write a
self-cancelling task more compactly like this:

task_scheduler.scheduleWithFixedDelay([this](){
    static int i = 0;
    logger.printfln("!!! %d", ++i);
    if (i >= 3)
        task_scheduler.cancel(task_scheduler.currentTask());
}, 1000, 1000);