task_scheduler: Use unique_ptr everywhere. Fix race condition.
Letting a task stay in the task queue while it is executed does not
work: To execute a task, we unlock the task queue mutex (so that the
task can schedule new tasks itself), while this happens, the task
and other FreeRTOS tasks (read as: preemptive threads, probably with
a higher priority) can create and remove tasks. After our task is done,
we don't know for sure that it is still the first entry in the queue.

This also renames currentTask to currentTaskId because it only returns
the ID of the current task, not a pointer to the task structure itself.