api, mqtt: Fix deadlock when calling commands.
a25b3f695 did not work as expected, because it blocked the MQTT thread.
If the main thread then tried to publish a message, it locks forever
because the MQTT thread has its own API lock.

The new approach is: callCommand with a ConfUpdate
(called only in the main thread) works without locking.
callCommand with a char * (called by API backends in other threads)
- locks the API's command mutex
- disables receiving new commands (by signalling the API backend
  to stop receiving messages)
- parses the command payload in a copy
- generates a task that (in the main thread) updates the config
  executes the callback and signals API backends to continue receiving
- and cancels an older task (for this API) that was already in flight

This makes sure that only one command is in flight, is efficient
(because one copy of the command config is necessary in any case)
and applies the _last_ received command if a burst happens faster than
we can process it. Intermediate commands can be dropped, but should
be idempotent anyway.