event_log: Fix writing unterminated strings.
When writing an unterminated string to the log, only the internal buffer
would respect the given string length. The length would be ignored when
writing to the serial port, usually resulting in binary garbage in the
console. The length is now respected when writing to the serial port
as well.