mqtt, http: Add _mqtt_thread and _HTTPThread overloads.
This allows to select whether the subscribe/on callbacks should be
executed in the main thread (more expensive, but can access config etc.
without task_scheduler.await) or in the MQTT/HTTP thread.

Especially mqtt.subscribe without _mqtt_thread is expensive now, because
we have to copy the payload into a separate buffer, as the MQTT thread
can not be blocked without deadlock.