web/config-component,-form: Refactor dirty handling
Remove dirty state from config-form saveDisabled. Rename it to
saveInProgress and just use it to avoid triggering another save
operation while a save operation is already in progress.

Move dirty state to config-component and allow to read it from
outside the actual config-form. This will be used to ignore update
to table rows in the new meters module.