Add new gethostbyname helper that ensure the result lifetime
The result is only valid of the call to the found_callback. This
doesn't allow for scheduling a task on the main thread that can
use the result.

The new helper always uses the found callback and schedules it on
the main thread and ensures the lifetime of the result.