cm_networking: Add separate thread to receive client messages.
If n clients sync up, a burst of n UDP packets is created. LWIP only
puts up to 6 packets (LWIP_UDP_RECVMBOX_SIZE) in the UDP receive buffer.
Further packets are dropped. If the order in a UDP burst is stable,
the dropped packets belong to the same charger(s) every time, resulting
in "unreachable charger" errors.

To fix this, add a separate FreeRTOS task that only receives packets
and puts them into a queue that is large enough to buffer one burst of
n packets (generated by n chargers).