form_row: Add explicit list of children using the ID context. Fixes #289.
FormRow did always generate a new ID and used this ID as the label's
for-attribute. However not every child uses the ID context resulting in
labels referencing unused IDs.

To fix this, add a list of child types known to use the ID context
and only generate and reference an ID if the passed child matches.

This is still not ideal, as we now get the same number of warnings
about labels without a nested input element or for-attribute. However
Chromium classifies this only as a "possible improvement", not a warning.