mqtt_auto_discovery: Fix external current minimum. Fixes #301.
Was limited to 6000 mA because 1 to 5999 mA are not allowed. However
other API users (for example EVCC) can write 0 mA to block charging.
Home Assistant then throws an error.

If a value between 1 and 5999 mA is now set via Home Assistant, the
value is ignored and reloading the page shows the correct value again.