energy_manager, power_manager: Try to preserve external phase request
If external control is enabled, assume that the last requested amount of phases
is whatever the contactor is switched to at the moment, in order to preserve
that across a reboot. This will still fail on a power cycle or bricklet update,
which set the contactor back to single phase.