meters-sun-spec: Avoid crash caused by connection error in the scan process
If a Modbus read does not succeed for a while then a reconnect is
triggered. If this happens while the common model block is being read
then an invalid pointer is freed. The pointer is made invalid at the
start of the read and made valid again at the end of the read. In the
crash case the reconnect happens while the pointer is invalid.

Avoid the whole pointer invalid making logic and just reread the common
model header to avoid the problem entirely.