wifi: Reorder AP config sequence
Both calls for setting the AP's SSID and IP enable the AP. Setting the IP
first will result in starting an open WiFi until the SSID is set. Instead,
set the SSID first to reduce the time the open WiFi is enabled. This
results in the AP having the correct SSID but having the wrong IP, but this
time frame is shorter than it takes for the first client to finish connecting.

Also, only print the amount of config attempts if it's not the expected 1.