web/charge_manager: Partly fix minimum/maximum current dependency
Automatically raising the maximum current with the minimum current
was broken when the maximum current was moved to another subpage.
Silently raising the maximum current on a currently invisible subpage
might not be a good idea either. Instead, limit the minimum current
to the maximum current. This limit is not enforced on save
if the minimum is already above the maximum or the maximum
is lowered below the minimum.