power_manager: Fix PV mode stopping a charge after 3p->1p switch
If the available power sharply reduces and the PM is stuck in 3p mode
for the full hysteresis time, the resulting power draw might have pushed
the filtered available power below zero. After switching from 3p to 1p,
the PM will not continue the charge if the available power appears to be
negative, resulting in a charge interruption for the full hysteresis time.
To avoid that, the cloud filter is reset after a 3p->1p switch so that
the available power matches the actual available power.