patches: Enforce a minimum TTL of 30s
This fixes the charge manager being unable to reach chargers
using certain host names.

The charge manager currently performs DNS lookups and packet sending in
separate iterations. If the TTL of a DNS reply is 0, it is discarded
right after the lookup and valid anymore by the time a packet should be
sent. If host names for chargers controlled by the charge manager have
a TTL of 0, the charge manager is unable to communicate with them.
Such host names are common for statically configured entries in HOSTS
files served by dnsmasq.