chucked-response: Ensure calls can only be made while they will be processed
The wait might timeout. No calls will be processed anymore. Now a late
call is made waiting forever for the have-result conditon to become true.
The call is then blocked, blocking the main thread, while holding the
ownership mutex, while the HTTP thread tries to take the mutex to clear
the ownership. Deadlock!