solar_forecast: Properly calculate next update times
To do this we schedule only when NTP is synced and afterwards
calculate the minimum next solar forecast API check time from
all planes.

The task scheduler is always run in interval mode, so it will
always automatically schedule an update if a new "next_update"
calculation can not be run.