http, web_server: Place read buffer on stack
The HTTP task's stack has to be large to execute handlers like the PDF generator.
When executing commands or responses, all work is done by other tasks and the
HTTP task's stack has enough room to fit the read buffer for those functions.
The HTTP task's stack actually has to be incread by 1k to fit the 4k buffer,
but this still saves 3k of global memory.