web/uplot: Draw vertical line text above chart bounding box
This means that a chart that uses vertical lines with text
needs to add an appropriate padding on top (with padding parameter).

It looks much better this way.

E.g. use padding height of 15 if you have a text height of 10 pixel
to get a space of 5 pixel between vertical line and text