charge_manager: Fix handling of EMv2
On a WEM, the CM cannot be enabled or disabled manually.
It was forced on only for EMv1. Now the enabled state on any WEM
will depend on whether or not any chargers are configured.
A EMv2 might be used only for heating, with no controlled chargers,
and the CM should be disabled there.