Rewrite NTP/RTC backend modules:
The RTC module is the central time keeping module. It updates attached
RTCs via IRtcBackend implementations whenever a new time is pushed.
NTP, OCPP and the (new) RTC Update API update the system time via
rtc.push_system_time. The RTC module makes sure that a high quality
time is not overwritten by a low quality time.

timestamp_minutes() and clock_synced() are now members of the RTC module,
which is why it is added as a dependency to a lot of modules.