config: Replace some problematic uses of asprintf
asprintf will place the generated strings on the heap, but heap memory
is not included in core dumps, which renders the generated error messages
useless.

Instead, error messages should be generated into buffers on the stack
because the whole stack will be included in a core dump. Message generation
should be moved to its own function that shouldn't be inlined, to avoid
bloating the calling function with the otherwise useless buffer
and the associated stack smashing protection.