evse_v2: Don't trigger button action during start-up
Pressing the button during the start-up phase will dispatch a trigger event
before the automation module has loaded its configuration, so it cannot be
processed. Drop the event instead, to remove the warning message in the log
during start-up. If the event was supposed to do something, the user can simply
press the button again.