config: Avoid additional string copies in error messages
The error might be caused by high memory pressure and an allocation
to copy the string is more likely to fail as well.

Instead, overwrite the zero-termination with the newline
for the log message and log the string as-is.

Also log StringBuilder overflows. A StringBuilder that happes to be filled
to the exact limit will incorrectly be considered overflowed as well,
but that is hopefully not that likely.