em-analysis: Retry persistent data loading if Bricklet wasn't ready
There was a race condition between the ESP32 asking the Bricklet for
the data and the Bricklet loading the data from the SD card. If the
ESP32 asked before the Bricket loaded the data then the ESP32 might
wrongly assume that the data is missing or corrupted, resulting in
dataloss. Now the Bricklet can report if the data was loaded already.
If not then the ESP32 will try again later.