em_energy_analysis: Fix calculating energy from power values
The power value for the last interval cannot be averaged with the new power value
because the update is only triggered on value change. Therefore, the old power value
was valid until just a moment ago, no matter how long the last interval actually was.

When calculating the duration in case of an overflow, another 1 must be added.
a - b = (a - b + 2^32) % 2^32 is true. The overflow is modulo 2^32, therefore 2^32
must be added, not just UINT32_MAX, which is 2^32-1. If we want to avoid the overflow
in the subtraction, this becomes a - b + 2^32 = (2^32-1) - b + a + 1. The compiler
will actually completely elide the if-else, calculating only a - b, which produces
the correct result even in case of an overflow.