web/meters: Compensate for download and parsing time of live and history
Assume that sending the HTTP request is reasonably quick and that
receiving and parsing the data might be slow. In this case the
actual true timestamp for the data is closer to the start of the
request than to the end of downloading and parsing the data. This
change should minimize the timestamping error in case of a slow
web browser.