web/meters: Don't trigger second cache init while first is running
Instead collect the new samples and append them as soon as the cache
initalization is done.

For this to work properly the previous change to take the timestamp
for the cache initalization before starting the download and parsing
is important. Otherwise the new samples might end up with a timestamp
older than the cache initalization resulting in wrong sample order.