web: Add simple C/C++ style preprocessor for module-available defines
Supported directives are #include, #define, #undef, #ifdef, #if, #else
and #endif.

In contrast to C/C++ a not defined symbol in an #if directive is an
error and does not evaluate to false.

For compatibilty with TypeScript all directives have to be prefixed
with //. Code removal by #if/#endif blocks is realized by commeting
out the TypeScript code using //. This preserves line numbers as no
lines are added or removed.

To make this work the base directory for the TypeScript import path
got moved from web/ to web/src/. This breaks absolute imports starting
with src/. They need to be changed to drop the src/.