api: Fix infinite recursion in deprecated addState function
Commit 6fb74cd30117146db973e94baa773d740c53816c added the new
keys_to_censor_in_debug_report parameter to addState, but also
left the addState call in the RTC module unchanged. The RTC
module called the deprecated setState function with true for
the interval_ms parameter, resulting in a compiler warning about
the bool-to-int comparison. The deprecated setState function
could not call the normal setState function anymore but called
itself in an infinite recursion resulting in a crash and a boot
loop.