day_ahead_prices: Improve get_[cheap|expensive]_hours_today function
We now calculate a proper starting index and just iterate from there
instead of searching through the whole data set for indices that
are in within today.

This is more efficient and we now can't accidentally add an index
that is out of range, since the range is predefined. With the old
approach it was possible go out-of-range when the day did not start
with index 0 (which was happening because of a bug in the
get_localtime_today_midnight_in_utc function)