day_ahead_prices: Remove prices.get call from get_*_15m functions
This is necessary for them to be callble from server.on_HTTPThread(...)
context. Additionally it makes sense to save the "first_date" for
the sorted prices so there can't be a race conditiopn where new prices
are received that have different first_date and the sorted prices are
not yet updated.