Extend generated enum comment support
A # comments are for the .enum file itself and don't show up in the
generated files.

A // comment at the start of a line is placed before the enum in the
generated header file.

A // comment after a enum value is placed after this eunm value in the
generated header file.