device_module: Fix reinitialization.
Creating a bindings device around a TF_TFP will mark the TFP context
as in use (by setting tfp->device). setup_device has to destroy the
device in case this is a reinitialization, to make sure the correct
TFP context is found. Contexts that are still in use (i.e. their
device was not destroyed) are not returned by tf_hal_get_tfp.