web_sockets: Kick out LRU client when a new one connects.
This is not ideal because if all connected clients are active,
we kick one out at random. If MAX_WEB_SOCKET_CLIENTS are connected
and a new one connects, the kicked out browser tab will reconnect
after some seconds and kick out another tab. But 6 more or less
usable tabs are better than 5 usable and a 6th one that is just
empty.

We could add a user-interaction timeout to the web interface, so that
it stops reconnecting if there was no input for (f.e.) 10 minutes.