config: Protect against creating "empty" configs
Don't use min = max = 0 for empty number config slots because that could be a valid configuration,
but the slot would be considered "empty". Use invalid min/max values instead and disallow creating
configs with those values.

Also disallow creating ConfUnions with a prototype nullptr, which would theoretically be possible,
because the slot would also be considered "empty" and it doesn't have a valid use case.