meters-sun-spec: Don't shrink register buffer to fix model 203
The detect_values function can choose to read fewer registers
than the initial amount. Before commit 72b4811ebefd572eac8b086
the initial buffer would have been reused and only filled with
fewer register values. Since that commit the initial buffer is
also shrunken to the new size. But the model validator might
access register values beyond the new size. This was okay before
but with the shrunken buffer the validator might access the
buffer out-of-bounds reading invalid random data. This stopped
the successful reading of models with this pattern. This affects
atleast model 203. Revert to the old behavior of not shrinking
the buffer.