ethernet: Fix disconnected network after temporary DHCP outage
Fix Ethernet staying in a disconnected state because of a missing GOT_IP event.

When a DHCP lease can't be renewed, a LOST_IP event is fired, but when the DHCP server
subsequently offers the same address as before, no GOT_IP fires because it only fires
on address change and the address did not change compared to before, in this case.
As a workaround, restart the DHCP client and clear the old address when a LOST_IP event
fires, so that the next DHCP reply will properly produce a GOT_IP event.