event_log: Download gzip-compressed trace log via remote access
The trace log can be huge and downloading it can time out over slow links.
As it is very redundant text, use gzip to compress it on the fly.
When accessing the web interface via LAN or WiFi, the uncompressed trace log
will be downloaded, but downloading it via remote access will use the compressed version.