iso15118: Add initial ISO15118 experimentations
In the long-run this module shall have the following features:

* SPI/QCA700x (Low level SPI communication with the QCA700x modem),
* SLAC (Signal-Level-Attenuation-Characterization) as described in ISO 15118-3,
* IPv6/ICMPv6/NDP (Neighbor Discorver Protoocol with Neighbor Solicitation/Advertisement),
* IPv6/UDP/SDP (SECC Discovery Protocol),
* IPv6/TCP/DIN-SPEC-70121,
* IPv6/TCP/ISO-15118-2,
* IPv6/TCP/ISO-15118-20

The EXI message encoding/decoding is based on libcbv2g / OpenV2G.

The objective is to *robustly* be able to
* fingerprint the EV for authentication,
* read SOC from the EV and
* if it will be implemented by EVs, allow for bi-directional charging.

To read the necessary data we may start a DC charge and end it again immediatily or similar, depending on the standards that the EV supports.
The actual charging will mostly still be done through IEC 61851-1 (PWM on CP).