wifi: Fix open AP when fallback mode is inactive after boot
Using WiFi.mode(...) to enable the AP when fallback mode is enabled
will immediately open an unencrypted AP until the correct security
settings are configured. If the fallback mode doesn't turn on,
the unencrypted AP will stay open forever.

Disconnecting the unencrypted AP immediately after enabling it
introduces a race condition. Instead, don't enable the AP by
default because it is enabled automatically when fallback mode
turns on.

The bug was introduced in cba0c27 and not released.