web: Improve Safari/WebKit detection for Firefox and Chrome on iOS
Firefox and Chrome on iOS also use Apple WebKit and suffer from the
digest auth problem that stops us from using a custom login dialog.
But before Firefox and Chrome on iOS would not be detected as such
correctly because they only have Safari/* in there user agent but not
Version/*. Use FxiOS/* and CriOS/* for detection also.