wifi: Remove obsolete softAPConfig loop
In the past, the soft-AP sometimes wouldn't accept its IP configuration.
As a workaround, setting the IP configuration was attempted in a loop until it was accepted.
This workaround doesn't seem to be neccessary anymore with the newer libraries.
Remove it because we don't want to have a possible infinite loop during start-up,
in case the IP configuration fails for a valid reason.