charge_manager: Fix current watchdog
Setting only available_current is not enough. The limits struct needs to be updated as well.
Use the available_current_update command to do it, which was already fixed for the automation action.

Also remove setting the max PV current via the API. If static charge management is used, all PV values should stay at 0.